chore(a2a-server): Refactor package structure.
