+++ packages/a2a-server/src/agent/task.test.ts
@@ -7,7 +7,7 @@
 import { describe, it, expect, vi } from 'vitest';
 import { Task } from './task.js';
 import type { Config, ToolCallRequestInfo } from '@google/gemini-cli-core';
-import { createMockConfig } from './testing_utils.js';
+import { createMockConfig } from '../utils/testing_utils.js';
 import type { ExecutionEventBus } from '@a2a-js/sdk/server';
 
 describe('Task', () => {
