+++ packages/a2a-server/src/index.ts
@@ -4,5 +4,6 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-export * from './agent.js';
+export * from './agent/executor.js';
+export * from './http/app.js';
 export * from './types.js';
