+++ packages/a2a-server/src/persistence/gcs.ts
@@ -13,12 +13,12 @@ import { tmpdir } from 'node:os';
 import { join } from 'node:path';
 import type { Task as SDKTask } from '@a2a-js/sdk';
 import type { TaskStore } from '@a2a-js/sdk/server';
-import { logger } from './logger.js';
-import { setTargetDir } from './config.js';
+import { logger } from '../utils/logger.js';
+import { setTargetDir } from '../config/config.js';
 import {
   getPersistedState,
   type PersistedTaskMetadata,
-} from './metadata_types.js';
+} from '../metadata_types.js';
 import { v4 as uuidv4 } from 'uuid';
 
 type ObjectType = 'metadata' | 'workspace';
